backend/requirements.txt
fastapi
uvicorn
python-dotenv
google-genai
google-auth
google-api-python-client
httpx
numpy
joblib
scikit-learn
scipy
firebase-admin
twilio
